Made routes/admin.js leaner by outsourcing all its content to a controller within controllers/admin.js
