+++ routes/admin.js
@@ -1,6 +1,7 @@
 const { restrict, checkAccess } = require("../lib/auth");
 const express = require("express");
 const router = express.Router();
+const multer = require("multer");
 const csrfProtection = csrf({ cookie: true });
 
 // Controller Imports
@@ -112,28 +113,19 @@ router.post(
 );
 
 // new menu item
-router.post("/admin/settings/menu/new", restrict, checkAccess, (req, res) => {
-  const result = newMenu(req);
-  if (result === false) {
-    res.status(400).json({ message: "Failed creating menu." });
-    return;
-  }
-  res.status(200).json({ message: "Menu created successfully." });
-});
+router.post(
+  "/admin/settings/menu/new",
+  restrict,
+  checkAccess,
+  adminController.postAdminSettingsNewMenuItemController
+);
 
 // update existing menu item
 router.post(
   "/admin/settings/menu/update",
   restrict,
   checkAccess,
-  (req, res) => {
-    const result = updateMenu(req);
-    if (result === false) {
-      res.status(400).json({ message: "Failed updating menu." });
-      return;
-    }
-    res.status(200).json({ message: "Menu updated successfully." });
-  }
+  adminController.postAdminSettingsUpdateMenuItemController
 );
 
 // delete menu item
@@ -141,14 +133,7 @@ router.post(
   "/admin/settings/menu/delete",
   restrict,
   checkAccess,
-  (req, res) => {
-    const result = deleteMenu(req, req.body.menuId);
-    if (result === false) {
-      res.status(400).json({ message: "Failed deleting menu." });
-      return;
-    }
-    res.status(200).json({ message: "Menu deleted successfully." });
-  }
+  adminController.postAdminSettingsDeleteMenuItemController
 );
 
 // We call this via a Ajax call to save the order from the sortable list
@@ -156,39 +141,14 @@ router.post(
   "/admin/settings/menu/saveOrder",
   restrict,
   checkAccess,
-  (req, res) => {
-    const result = orderMenu(req, res);
-    if (result === false) {
-      res.status(400).json({ message: "Failed saving menu order" });
-      return;
-    }
-    res.status(200).json({});
-  }
+  adminController.postAdminSettingsMenuSaveOrderController
 );
 
 // validate the permalink
-router.post("/admin/validatePermalink", async (req, res) => {
-  // if doc id is provided it checks for permalink in any products other that one provided,
-  // else it just checks for any products with that permalink
-  const db = req.app.db;
-
-  let query = {};
-  if (typeof req.body.docId === "undefined" || req.body.docId === "") {
-    query = { productPermalink: req.body.permalink };
-  } else {
-    query = {
-      productPermalink: req.body.permalink,
-      _id: { $ne: getId(req.body.docId) },
-    };
-  }
-
-  const products = await db.products.countDocuments(query);
-  if (products && products > 0) {
-    res.status(400).json({ message: "Permalink already exists" });
-    return;
-  }
-  res.status(200).json({ message: "Permalink validated successfully" });
-});
+router.post(
+  "/admin/validatePermalink",
+  adminController.postAdminValidatePermanentLinkController
+);
 
 // Discount codes
 router.get(
@@ -196,23 +156,7 @@ router.get(
   csrfProtection,
   restrict,
   checkAccess,
-  async (req, res) => {
-    const db = req.app.db;
-
-    const discounts = await db.discounts.find({}).toArray();
-
-    res.render("settings-discounts", {
-      title: "Discount code",
-      config: req.app.config,
-      session: req.session,
-      discounts,
-      admin: true,
-      message: clearSessionValue(req.session, "message"),
-      messageType: clearSessionValue(req.session, "messageType"),
-      helpers: req.handlebars.helpers,
-      csrfToken: req.csrfToken(),
-    });
-  }
+  adminController.getAdminSettingsDiscountsController
 );
 
 // Edit a discount code
@@ -221,23 +165,7 @@ router.get(
   csrfProtection,
   restrict,
   checkAccess,
-  async (req, res) => {
-    const db = req.app.db;
-
-    const discount = await db.discounts.findOne({ _id: getId(req.params.id) });
-
-    res.render("settings-discount-edit", {
-      title: "Discount code edit",
-      session: req.session,
-      admin: true,
-      discount,
-      message: clearSessionValue(req.session, "message"),
-      messageType: clearSessionValue(req.session, "messageType"),
-      helpers: req.handlebars.helpers,
-      config: req.app.config,
-      csrfToken: req.csrfToken(),
-    });
-  }
+  adminController.getAdminSettingsDiscountEditController
 );
 
 // Update discount code
@@ -245,68 +173,7 @@ router.post(
   "/admin/settings/discount/update",
   restrict,
   checkAccess,
-  async (req, res) => {
-    const db = req.app.db;
-
-    // Doc to insert
-    const discountDoc = {
-      discountId: req.body.discountId,
-      code: req.body.code,
-      type: req.body.type,
-      value: parseInt(req.body.value),
-      start: moment(req.body.start, "DD/MM/YYYY HH:mm").toDate(),
-      end: moment(req.body.end, "DD/MM/YYYY HH:mm").toDate(),
-    };
-
-    // Validate the body again schema
-    const schemaValidate = validateJson("editDiscount", discountDoc);
-    if (!schemaValidate.result) {
-      res.status(400).json(schemaValidate.errors);
-      return;
-    }
-
-    // Check start is after today
-    if (moment(discountDoc.start).isBefore(moment())) {
-      res
-        .status(400)
-        .json({ message: "Discount start date needs to be after today" });
-      return;
-    }
-
-    // Check end is after the start
-    if (!moment(discountDoc.end).isAfter(moment(discountDoc.start))) {
-      res
-        .status(400)
-        .json({ message: "Discount end date needs to be after start date" });
-      return;
-    }
-
-    // Check if code exists
-    const checkCode = await db.discounts.countDocuments({
-      code: discountDoc.code,
-      _id: { $ne: getId(discountDoc.discountId) },
-    });
-    if (checkCode) {
-      res.status(400).json({ message: "Discount code already exists" });
-      return;
-    }
-
-    // Remove discountID
-    delete discountDoc.discountId;
-
-    try {
-      await db.discounts.updateOne(
-        { _id: getId(req.body.discountId) },
-        { $set: discountDoc },
-        {}
-      );
-      res
-        .status(200)
-        .json({ message: "Successfully saved", discount: discountDoc });
-    } catch (ex) {
-      res.status(400).json({ message: "Failed to save. Please try again" });
-    }
-  }
+  adminController.postAdminSettingsUpdateDiscountCodeController
 );
 
 // Create a discount code
@@ -315,18 +182,7 @@ router.get(
   csrfProtection,
   restrict,
   checkAccess,
-  async (req, res) => {
-    res.render("settings-discount-new", {
-      title: "Discount code create",
-      session: req.session,
-      admin: true,
-      message: clearSessionValue(req.session, "message"),
-      messageType: clearSessionValue(req.session, "messageType"),
-      helpers: req.handlebars.helpers,
-      config: req.app.config,
-      csrfToken: req.csrfToken(),
-    });
-  }
+  adminController.getAdminSettingsNewDiscountCode
 );
 
 // Create a discount code
@@ -335,57 +191,7 @@ router.post(
   csrfProtection,
   restrict,
   checkAccess,
-  async (req, res) => {
-    const db = req.app.db;
-
-    // Doc to insert
-    const discountDoc = {
-      code: req.body.code,
-      type: req.body.type,
-      value: parseInt(req.body.value),
-      start: moment(req.body.start, "DD/MM/YYYY HH:mm").toDate(),
-      end: moment(req.body.end, "DD/MM/YYYY HH:mm").toDate(),
-    };
-
-    // Validate the body again schema
-    const schemaValidate = validateJson("newDiscount", discountDoc);
-    if (!schemaValidate.result) {
-      res.status(400).json(schemaValidate.errors);
-      return;
-    }
-
-    // Check if code exists
-    const checkCode = await db.discounts.countDocuments({
-      code: discountDoc.code,
-    });
-    if (checkCode) {
-      res.status(400).json({ message: "Discount code already exists" });
-      return;
-    }
-
-    // Check start is after today
-    if (moment(discountDoc.start).isBefore(moment())) {
-      res
-        .status(400)
-        .json({ message: "Discount start date needs to be after today" });
-      return;
-    }
-
-    // Check end is after the start
-    if (!moment(discountDoc.end).isAfter(moment(discountDoc.start))) {
-      res
-        .status(400)
-        .json({ message: "Discount end date needs to be after start date" });
-      return;
-    }
-
-    // Insert discount code
-    const discount = await db.discounts.insertOne(discountDoc);
-    res.status(200).json({
-      message: "Discount code created successfully",
-      discountId: discount.insertedId,
-    });
-  }
+  adminController.getAdminSettingsNewDiscountCodeController
 );
 
 // Delete discount code
@@ -393,19 +199,7 @@ router.delete(
   "/admin/settings/discount/delete",
   restrict,
   checkAccess,
-  async (req, res) => {
-    const db = req.app.db;
-
-    try {
-      await db.discounts.deleteOne({ _id: getId(req.body.discountId) }, {});
-      res.status(200).json({ message: "Discount code successfully deleted" });
-      return;
-    } catch (ex) {
-      res
-        .status(400)
-        .json({ message: "Error deleting discount code. Please try again." });
-    }
-  }
+  adminController.deleteAdminSettingsDeleteDiscountCodeController
 );
 
 // upload the file
@@ -415,209 +209,20 @@ router.post(
   restrict,
   checkAccess,
   upload.single("uploadFile"),
-  async (req, res) => {
-    const db = req.app.db;
-
-    if (req.file) {
-      const file = req.file;
-
-      // Get the mime type of the file
-      const mimeType = mime.lookup(file.originalname);
-
-      // Check for allowed mime type and file size
-      if (!allowedMimeType.includes(mimeType) || file.size > fileSizeLimit) {
-        // Remove temp file
-        fs.unlinkSync(file.path);
-
-        // Return error
-        res.status(400).json({
-          message: "File type not allowed or too large. Please try again.",
-        });
-        return;
-      }
-
-      // get the product form the DB
-      const product = await db.products.findOne({
-        _id: getId(req.body.productId),
-      });
-      if (!product) {
-        // delete the temp file.
-        fs.unlinkSync(file.path);
-
-        // Return error
-        res
-          .status(400)
-          .json({ message: "File upload error. Please try again." });
-        return;
-      }
-
-      const productPath = product._id.toString();
-      const uploadDir = path.join("public/uploads", productPath);
-
-      // Check directory and create (if needed)
-      checkDirectorySync(uploadDir);
-
-      // Setup the new path
-      const imagePath = path.join(
-        "/uploads",
-        productPath,
-        file.originalname.replace(/ /g, "_")
-      );
-
-      // save the new file
-      const dest = fs.createWriteStream(
-        path.join(uploadDir, file.originalname.replace(/ /g, "_"))
-      );
-      const pipeline = util.promisify(stream.pipeline);
-
-      try {
-        await pipeline(fs.createReadStream(file.path), dest);
-
-        // delete the temp file.
-        fs.unlinkSync(file.path);
-
-        // if there isn't a product featured image, set this one
-        if (!product.productImage) {
-          await db.products.updateOne(
-            { _id: getId(req.body.productId) },
-            { $set: { productImage: imagePath } },
-            { multi: false }
-          );
-        }
-        res.status(200).json({ message: "File uploaded successfully" });
-      } catch (ex) {
-        console.log("Failed to upload the file", ex);
-        res
-          .status(400)
-          .json({ message: "File upload error. Please try again." });
-      }
-    } else {
-      // Return error
-      console.log("fail", req.file);
-      res.status(400).json({ message: "File upload error. Please try again." });
-    }
-  }
+  adminController.postAdminFileUploadController
 );
 
 // delete a file via ajax request
-router.post("/admin/testEmail", restrict, (req, res) => {
-  const config = req.app.config;
-  // TODO: Should fix this to properly handle result
-  sendEmail(
-    config.emailAddress,
-    "expressCart test email",
-    "Your email settings are working"
-  );
-  res.status(200).json({ message: "Test email sent" });
-});
+router.post(
+  "/admin/testEmail",
+  restrict,
+  adminController.postAdminTestEmailController
+);
 
-router.post("/admin/searchall", restrict, async (req, res, next) => {
-  const db = req.app.db;
-  const searchValue = req.body.searchValue;
-  const limitReturned = 5;
-
-  // Empty arrays
-  let customers = [];
-  let orders = [];
-  let products = [];
-
-  // Default queries
-  const customerQuery = {};
-  const orderQuery = {};
-  const productQuery = {};
-
-  // If an ObjectId is detected use that
-  if (ObjectId.isValid(req.body.searchValue)) {
-    // Get customers
-    customers = await db.customers
-      .find({
-        _id: ObjectId(searchValue),
-      })
-      .limit(limitReturned)
-      .sort({ created: 1 })
-      .toArray();
-
-    // Get orders
-    orders = await db.orders
-      .find({
-        _id: ObjectId(searchValue),
-      })
-      .limit(limitReturned)
-      .sort({ orderDate: 1 })
-      .toArray();
-
-    // Get products
-    products = await db.products
-      .find({
-        _id: ObjectId(searchValue),
-      })
-      .limit(limitReturned)
-      .sort({ productAddedDate: 1 })
-      .toArray();
-
-    return res.status(200).json({
-      customers,
-      orders,
-      products,
-    });
-  }
-
-  // If email address is detected
-  if (emailRegex.test(req.body.searchValue)) {
-    customerQuery.email = searchValue;
-    orderQuery.orderEmail = searchValue;
-  } else if (numericRegex.test(req.body.searchValue)) {
-    // If a numeric value is detected
-    orderQuery.amount = req.body.searchValue;
-    productQuery.productPrice = req.body.searchValue;
-  } else {
-    // String searches
-    customerQuery.$or = [
-      { firstName: { $regex: new RegExp(searchValue, "img") } },
-      { lastName: { $regex: new RegExp(searchValue, "img") } },
-    ];
-    orderQuery.$or = [
-      { orderFirstname: { $regex: new RegExp(searchValue, "img") } },
-      { orderLastname: { $regex: new RegExp(searchValue, "img") } },
-    ];
-    productQuery.$or = [
-      { productTitle: { $regex: new RegExp(searchValue, "img") } },
-      { productDescription: { $regex: new RegExp(searchValue, "img") } },
-    ];
-  }
-
-  // Get customers
-  if (Object.keys(customerQuery).length > 0) {
-    customers = await db.customers
-      .find(customerQuery)
-      .limit(limitReturned)
-      .sort({ created: 1 })
-      .toArray();
-  }
-
-  // Get orders
-  if (Object.keys(orderQuery).length > 0) {
-    orders = await db.orders
-      .find(orderQuery)
-      .limit(limitReturned)
-      .sort({ orderDate: 1 })
-      .toArray();
-  }
-
-  // Get products
-  if (Object.keys(productQuery).length > 0) {
-    products = await db.products
-      .find(productQuery)
-      .limit(limitReturned)
-      .sort({ productAddedDate: 1 })
-      .toArray();
-  }
-
-  return res.status(200).json({
-    customers,
-    orders,
-    products,
-  });
-});
+router.post(
+  "/admin/searchall",
+  restrict,
+  adminController.postAdminSearchAllController
+);
 
 module.exports = router;
